% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_score.R
\name{brier}
\alias{brier}
\title{Brier Score
\code{brier()} returns the Brier score obtained from a given prediction and known result}
\usage{
brier(act, pred)
}
\arguments{
\item{act}{actual results}

\item{pred}{predicted probabilties}
}
\value{
numeric.
}
\description{
Brier Score
\code{brier()} returns the Brier score obtained from a given prediction and known result
}
