% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_dcapply.R
\name{dcapply}
\alias{dcapply}
\title{Do Call Apply}
\usage{
dcapply(x, fun, combine, cores = 1, ...)
}
\arguments{
\item{x}{list. the input list}

\item{fun}{function. the function to map over \code{x}}

\item{combine}{function. the function to call on the mapped results}

\item{cores}{numeric. Number of cores to use in parallel. Default \code{1}.}

\item{...}{arguments passed to \code{fun}.}
}
\value{
a single entity.
}
\description{
\code{dcapply()} uses \code{do.call()} to merge the products of an applied function according to specifications
}
\details{
This function is no longer used for anything in \code{dryscrape}.

The function will be applied in parallel if \code{cores >= 1}.
}
