% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ftable.R
\name{ftable2df}
\alias{ftable2df}
\title{F Table to Data Frame
\code{ftable2df()} returns a \code{data.frame} from an \code{ftable} object}
\usage{
ftable2df(mydata)
}
\arguments{
\item{mydata}{an \code{ftable}}
}
\value{
a \code{data.frame}
}
\description{
F Table to Data Frame
\code{ftable2df()} returns a \code{data.frame} from an \code{ftable} object
}
