% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_score.R
\name{log_loss}
\alias{log_loss}
\title{Logarithmic Loss}
\usage{
log_loss(act, pred, allow_inf = FALSE)
}
\arguments{
\item{act}{actual results}

\item{pred}{predicted probabilties}

\item{allow_inf}{pass through infinite loss? \code{FALSE} replaces Inf with large but finite penalties at the extremes. Default \code{FALSE}.}
}
\value{
numeric.
}
\description{
\code{log_loss()} returns the logarithmic loss obtained from a given prediction and known result
}
\details{
The allow_inf parameter controls whether infinite loss is allowed (default is FALSE)
Setting allow_inf to FALSE will cause large but finite penalties at the extremes
}
