% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_one_liners.R
\name{moving}
\alias{moving}
\title{Moving
\code{moving()} returns a vector of averages obtained from the n elements of x preceding and including the element at each respective index}
\usage{
moving(x, n = 5)
}
\arguments{
\item{x}{the input vector.}

\item{n}{the window size. Default \code{5}.}
}
\value{
a vector of moving averages.
}
\description{
Moving
\code{moving()} returns a vector of averages obtained from the n elements of x preceding and including the element at each respective index
}
