% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_one_liners.R
\name{na_as_string}
\alias{na_as_string}
\title{NA as String
\code{na_as_string()} returns a character vector with \code{NA} values aliased to \code{"NA"}}
\usage{
na_as_string(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
character vector.
}
\description{
NA as String
\code{na_as_string()} returns a character vector with \code{NA} values aliased to \code{"NA"}
}
