% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_one_liners.R
\name{na_as_zero}
\alias{na_as_zero}
\title{NA as zero
\code{na_as_zero()} returns a numeric vector with \code{NA} values aliased to \code{0}}
\usage{
na_as_zero(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
numeric vector.
}
\description{
NA as zero
\code{na_as_zero()} returns a numeric vector with \code{NA} values aliased to \code{0}
}
