% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_one_liners.R
\name{na_if_null}
\alias{na_if_null}
\title{NA if NULL
\code{na_if_null()} returns an object's value if it is not \code{NULL} and \code{NA} otherwise}
\usage{
na_if_null(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
\code{x} unless \code{x} is \code{NULL}, in which case \code{NA}
}
\description{
NA if NULL
\code{na_if_null()} returns an object's value if it is not \code{NULL} and \code{NA} otherwise
}
