% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_one_liners.R
\name{nabs}
\alias{nabs}
\title{Numeric Absolute}
\usage{
nabs(x)
}
\arguments{
\item{x}{vector.}
}
\value{
numeric vector.
}
\description{
\code{nabs()} returns x after first converting it to class numeric via character
}
\details{
Its primary use is converting objects of class factor to numeric
}
